requirements.txt
fastapi==0.85.0
uvicorn==0.20.0
pillow==9.3.0
numpy==1.23.5
tensorflow==2.11.0
